requirements.txt
Flask
Flask-Login
Flask-Mail
Flask-SQLAlchemy
itsdangerous
numpy
faiss-cpu
transformers
torch
Faker
google-generativeai
python-dotenv
